Separate SqlServer code in optimistic concurrency tests
The optimistic concurrency test base classes contained some
SqlServer-specific code, i.e. the setup of a byte[] rowversion as the
concurrency column. Moved that out to the SqlServer tests.
